+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Consistent use of try-with-resources when handling streams
 * Updated AbstractRequest to remove deprecated methods uploadFileItems and readBytes (were protected static). Use StreamUtils instead.
 * Replaced org.apache.tika:tika library with org.overviewproject:mime-types in FileUtil to validate uploaded file mime types.
+* Updated FileUtil to make file extension and mime type validation case insensitive.
 
 ### Bug Fixes